Added human check if x is pressed
If you press x the camera takes a picture and feeds it into a TensorFlow object recognition to check if there is a human or not. I tried to add just the function but couldn't get it working, so I just copied the hole script onto the x press. Whenever I figure out how to do that I'll fix it